Re-Organization
Moved SettingSectionKeys to its own class so it can be used by add-on plugins

Moved IllegalWorldException to new exceptions package
