+++ src/main/java/org/shanerx/tradeshop/objects/ShopChest.java
@@ -34,9 +34,9 @@ import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.persistence.PersistentDataType;
-import org.shanerx.tradeshop.IllegalWorldException;
 import org.shanerx.tradeshop.TradeShop;
 import org.shanerx.tradeshop.enumys.DebugLevels;
+import org.shanerx.tradeshop.exceptions.IllegalWorldException;
 import org.shanerx.tradeshop.utils.Utils;
 
 import java.util.HashMap;
